releasing SS2 with loom bug fix (#105)
* Adding a note about haplotype caller version (#95)

* Adding a note about haplotype caller version

In this pipeline, by default the variant calling is done with GATK3.5 unless the user specifies an input parameter to use GATK4. We don't explain that anywhere and our example inputs use the defaults. 

I would suggest to put more documentation / example inputs but my hope is we will be going to DRAGEN-GATK genome pipeline soon and this will be deprecated so maybe not worth it. 

Other side question - I thought we updated documentation about the cromwell version support but at least develop seems to still note it was successfully tested on cromwell 47 - should we say it's now tested on cromwell 52 and newer? @ekiernan

* Update README.md

* Update README.md

Co-authored-by: ekiernan <55763654+ekiernan@users.noreply.github.com>

* changed metadata string to a list (#103)

* changed metadata string to a list

* version update

* reverted docker change for Optimus

* Update pipelines/skylab/smartseq2_multisample/MultiSampleSmartSeq2.changelog.md

Co-authored-by: Kylee Degatano <kdegatano@broadinstitute.org>

* Update pipelines/skylab/smartseq2_single_sample/SmartSeq2SingleSample.changelog.md

Co-authored-by: Kylee Degatano <kdegatano@broadinstitute.org>

* Changed Optimus version

Co-authored-by: Kylee Degatano <kdegatano@broadinstitute.org>

Co-authored-by: Kylee Degatano <kdegatano@broadinstitute.org>
Co-authored-by: ekiernan <55763654+ekiernan@users.noreply.github.com>