Update warp-tools docker version
Updated the warp-tools docker version in FastqProcessing.wdl and Metrics.wdl. 

This update was required due to bug fix in a recent pull request in warp-tools (https://github.com/broadinstitute/warp-tools/pull/18). The issue was integer division that caused the percent of mitochondrial molecules to always be calculated as zero.

As a result, a new version of the warp-tools docker was built and pushed into gcr. 

----

### Checklist 
If you can answer "yes" to the following items, please add a checkmark next to the appropriate checklist item(s) **and** notify our WARP documentation team by tagging either @ekiernan or @kayleemathews in a comment on this PR.

- [ ] Did you add inputs, outputs, or tasks to a workflow?
- [x] Did you modify, delete or move: file paths, file names, input names, output names, or task names?
- [ ] If you made a changelog update, did you update the pipeline version number?