Update Puck_210817_11.mm10.json
Updated the json file with the appropriate inputs. Parameters were based on an older version of the pipeline. Thus, they needed to be updated to match that of the current pipeline version. The whitelist parameter, for example, was removed as it is no longer found in the newer version of pipeline. All files were copied into broad-gotc-test-storage and file links were edited accordingly.