Updating feature branch with develop (#928)
* Add input_id to cell metrics and gene metrics outputs of Optimus (#912)

* name gene and cell metrics using input_id

* changelog updates

* rename output file

* Lk pd 2134 optimus chem (#914)

Modified the Optimus chemistry input; by default, 10x v2 or v3 whitelists will be selected based on the new chemistry input.

* Update warp-tools docker version

Updated the warp-tools docker version in FastqProcessing.wdl and Metrics.wdl. 

This update was required due to bug fix in a recent pull request in warp-tools (https://github.com/broadinstitute/warp-tools/pull/18). The issue was integer division that caused the percent of mitochondrial molecules to always be calculated as zero.

As a result, a new version of the warp-tools docker was built and pushed into gcr. 

----

### Checklist 
If you can answer "yes" to the following items, please add a checkmark next to the appropriate checklist item(s) **and** notify our WARP documentation team by tagging either @ekiernan or @kayleemathews in a comment on this PR.

- [ ] Did you add inputs, outputs, or tasks to a workflow?
- [x] Did you modify, delete or move: file paths, file names, input names, output names, or task names?
- [ ] If you made a changelog update, did you update the pipeline version number?

* Update Metrics.wdl

Updated the docker version.

* Update Optimus.changelog.md (#920)

* install matplotlib<3.7 to fix scanpy (#921)

* install matplotlib<3.7 to fix scanpy

* need quotes

* Update Optimus.changelog.md

* dynamically size disk in checkOptimusInput (#924)

* dynamically size disk in checkOptimusInput

* changelog updates

---------

Co-authored-by: Nikelle Petrillo <38223776+nikellepetrillo@users.noreply.github.com>
Co-authored-by: ekiernan <55763654+ekiernan@users.noreply.github.com>
Co-authored-by: aawdeh <awdeh@broadinstitute.org>