Create atac.wdl
These two tasks (TrimAdapters and BWAPairedEndAlignment) comprise the first step of ATAC portion of the Multiome pipeline. 

snapATAC2 will be used for the preprocessing of scATAC-seq. However, since snapATAC2 requires an aligned bam as input, this alignment step is needed prior to running snapATAC2. 

The two tasks were isolated from this WDL: https://github.com/broadinstitute/warp/blob/develop/beta-pipelines/skylab/ATAC/ATAC.wdl. They were modified to allow the user to select more options as input, such as disk and memory sizes for each task, and the number of threads needed for BWAPairedEndAlignment. 

If the user does not manually input disk size, the program will dynamically allocate disk space for the task. Future work will include applying the same logic for memory and number threads. 

The BWA version used is 0.7.17. 

With the default values for memory, disk and number of threads for each of the tasks, this script can run on a pair of compressed FASTQ files of size 5GB each.